Fixing imports and combining views
