+++ app/__init__.py
@@ -1,7 +1,7 @@
 from flask import Flask, Blueprint
+from app.api.v1 import version1
 
 def create_app():
     app = Flask(__name__)
-    from .api.v1 import version1 as v1
-    app.register_blueprint(v1)
+    app.register_blueprint(version1)
     return app
